// src/api/endpoints/model/createSysMenuRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * (title)
 * OpenAPI spec version: 0.0.0
 */
import type { SysMenuType } from "./sysMenuType";

export interface CreateSysMenuRequest {
  parentId?: string;
  title: string;
  type: SysMenuType;
  path?: string;
  component?: string;
  perms?: string;
  icon?: string;
  sortOrder?: number;
}